💃 v0.0.3: Complete code refactoring, file structuring, and fixed major issue #4
