+++ .gitignore
@@ -39,7 +39,8 @@ bower_components
 build/Release
 
 # Dependency directories
-node_modules/
+node_modules/*
+!node_modules/BulkMailCli_*/
 jspm_packages/
 
 # TypeScript v1 declaration files
