+++ index.js
@@ -1,0 +1,4 @@
+// This file is just used for exporting PROJECT_DIR variable.
+module.exports = {
+    PROJECT_DIR : __dirname
+}
