+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bulk-mail-cli",
-  "version": "0.0.2",
+  "version": "0.0.3",
   "description": "Send bulk non-spammy emails right from your terminal!",
   "main": "src/dist/src/bulkmailer.js",
   "scripts": {
@@ -24,6 +24,9 @@
     "babel-preset-env": "^1.7.0",
     "rimraf": "^2.6.2"
   },
+  "bundledDependencies": [
+    "BulkMailCli_settings"
+  ],
   "bin": {
     "bulkmail": "./src/dist/src/bin/global.js"
   },
