+++ src/utils/commands/commands.class.js
@@ -1,6 +1,16 @@
+/**
+ * @class @name BulkMailCli_commands
+ * @type class
+ * @description Does different things for different bulkmail commands
+ * 
+ * Date of creation: Sun, 4th Nov 2018. 10:43:17 IST
+ */
+
+
 import BulkMailCli_minimist from '../minimist.util'
 import { terminal } from 'terminal-kit'
 import help from './tools/help.util'
+
 
 class BulkMailCli_commands {
